Back-end-php
You can write the name of a Brazilian city and you'll be redirected to data about covid in those cities. I used PHP and HTML to receive the input, look for it on a database and than display the information fetched. ( Unfortunately the I couldn't upload the file I used to get the information from, but it was all provided by IBGE.